+++ pajbot/apiwrappers/seventv/__init__.py
@@ -1,1 +1,5 @@
 from .api import SevenTVAPI
+
+__all__ = [
+    "SevenTVAPI",
+]