+++ pajbot/web/common/__init__.py
@@ -1,3 +1,3 @@
-import pajbot.web.common.assets
-import pajbot.web.common.filters
-import pajbot.web.common.menu
+from . import assets as assets
+from . import filters as filters
+from . import menu as menu