+++ main.py
@@ -58,7 +58,7 @@ def handle_exceptions(exctype, value, tb):
 
 
 if __name__ == "__main__":
-    from pajbot.utils import init_logging, dump_threads
+    from pajbot.utils import dump_threads, init_logging
 
     def on_sigusr1(signal, frame):
         log.info("Process was interrupted with SIGUSR1, dumping all thread stack traces")