+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` w
 - Dev: Add typing to the Sub Alert module. (#2512)
 - Dev: Add typing to the Raid Alert module. (#2513)
 - Dev: Add typing to the remaining chat alert modules. (#2514)
+- Dev: Add more typing to the CLR Overlay modules. (#2530)
 
 ## v1.67