+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` w
 - Dev: Add typing to the Schedule manager. (#2531)
 - Dev: Add more typing to the Command Manager. (#2532)
 - Dev: Change to nkgilleys' flask-assets fork to add Flask3 support. (#2571)
+- Dev: Add experimental gunicorn support. (#2572)
 
 ## v1.67